Complete MOV mechanics (no mov extend) for the x86-64 emitter using the new interface

+++ compiler/ecpps/src/CodeGeneration/Emitters/x86_64.cpp
@@ -393,8 +393,7 @@ struct ecpps::codegen::emitters::EmitSpecificMovImpl<ecpps::codegen::emitters::O
                return x86_64::GenerateMovImmToMem32(destinationRegister, destinationDisplacement,
                                                     static_cast<std::uint32_t>(sourceImmediate));
           case ecpps::abi::qwordSize:
-               return x86_64::GenerateMovImmToMem64(destinationRegister, destinationDisplacement,
-                                                    static_cast<std::uint32_t>(sourceImmediate));
+               return x86_64::GenerateMovImmToMem64(destinationRegister, destinationDisplacement, sourceImmediate);
           default: throw TracedException(std::logic_error("Invalid mov operation"));
           }
      }
